refactor : Sort component
타 페이지에서 사용해야하는 Sort 컴포넌트 파일 리팩토링

Related to : #42
