+++ src/app/best/page.tsx
@@ -1,5 +1,6 @@
-import Category from '@/components/best/Category';
+import Category from '@/components/best/category/Category';
 import PageHeader from '@/components/common/PageHeader';
+import Sort from '@/components/common/Sort';
 import styles from '@/styles/best/best.module.css';
 
 export default function bestPage() {
@@ -9,13 +10,7 @@ export default function bestPage() {
 			<div className={styles.container}>
 				<div className={styles.wrapper}>
 					<Category />
-					<div className={styles.sortBox}>
-						<div className={styles.productAmounts}>상품 (123)</div>
-						<div className={styles.sortItem}>
-							<div>인기순</div>
-							<div>30</div>
-						</div>
-					</div>
+					<Sort />
 				</div>
 			</div>
 		</>
