+++ src/styles/best/best.module.css
@@ -18,45 +18,3 @@
 .wrapper > :last-child {
 	margin-bottom: 100px;
 }
-
-/* sort */
-.sortBox {
-	width: inherit;
-	height: 50px;
-	margin-top: -50px;
-
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-
-	border: 1px solid #aaaaaa;
-}
-.productAmounts {
-	width: 150px;
-	height: 36px;
-	margin-left: 20px;
-	font-size: 20px;
-
-	display: flex;
-	align-items: center;
-}
-.sortItem {
-	width: 220px;
-	margin-right: 20px;
-
-	display: flex;
-	justify-content: space-between;
-}
-.sortItem > div {
-	width: 100px;
-	height: 36px;
-	border-radius: 10px;
-	font-size: 18px;
-
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	cursor: pointer;
-
-	border: 1px solid #aaaaaa;
-}
